move `ColDesc` to `cassandra.util` to avoid circular dependecies
since our `UnixSocketEndPoint` was imported from polices, that was
causing a circular dependecy, moving `ColDesc` out of `cassandra.polices`
fixes that issue.
