+++ tests/integration/standard/column_encryption/test_policies.py
@@ -17,7 +17,7 @@ import unittest
 
 from tests.integration import use_singledc, TestCluster
 
-from cassandra.policies import ColDesc
+from cassandra.util import ColDesc
 
 from cassandra.column_encryption.policies import AES256ColumnEncryptionPolicy, \
     AES256_KEY_SIZE_BYTES, AES256_BLOCK_SIZE_BYTES
